Stuff starts to has shape, and it compiles
